+++ src/api/modules/competitions.ts
