+++ src/commands/instances/group/GroupRecords.ts
@@ -1,26 +1,26 @@
 import { EmbedFieldData, MessageEmbed } from 'discord.js';
-import { fetchGroupDetails, fetchGroupGained } from '../../../api/modules/group';
-import { GroupGainedEntry } from '../../../api/types';
+import { fetchGroupDetails, fetchGroupRecords } from '../../../api/modules/groups';
+import { GroupRecordEntry } from '../../../api/types';
 import config from '../../../config';
 import { Command, ParsedMessage } from '../../../types';
-import { getEmoji, getMetricName, toKMB } from '../../../utils';
+import { formatDate, getEmoji, getMetricName, toKMB } from '../../../utils';
 import CommandError from '../../CommandError';
 
-class GainedCommand implements Command {
+class RecordsCommand implements Command {
   name: string;
   template: string;
   requiresAdmin?: boolean | undefined;
   requiresGroup?: boolean | undefined;
 
   constructor() {
-    this.name = 'View group gains';
-    this.template = '!group gained {period} {metric}';
+    this.name = 'View group records';
+    this.template = '!group records {period} {metric}';
     this.requiresGroup = true;
   }
 
   activated(message: ParsedMessage) {
     const { command, args } = message;
-    return command === 'group' && args.length >= 2 && args[0] === 'gained';
+    return command === 'group' && args.length >= 2 && args[0] === 'records';
   }
 
   async execute(message: ParsedMessage) {
@@ -30,14 +30,14 @@ class GainedCommand implements Command {
 
     try {
       const group = await fetchGroupDetails(groupId);
-      const gained = await fetchGroupGained(groupId, period, metric);
-      const pageURL = `https://wiseoldman.net/groups/${groupId}/gained/`;
-      const fields = this.buildGainedFields(gained);
+      const records = await fetchGroupRecords(groupId, period, metric);
+      const pageURL = `https://wiseoldman.net/groups/${groupId}/records/`;
+      const fields = this.buildRecordFields(records);
       const icon = getEmoji(metric);
 
       const response = new MessageEmbed()
         .setColor(config.visuals.blue)
-        .setTitle(`${icon} ${group.name} ${getMetricName(metric)} gains (${period})`)
+        .setTitle(`${icon} ${group.name} ${getMetricName(metric)} records (${period})`)
         .setURL(pageURL)
         .addFields(fields);
 
@@ -47,18 +47,19 @@ class GainedCommand implements Command {
     }
   }
 
-  buildGainedFields(gained: GroupGainedEntry[]): EmbedFieldData[] {
-    return gained.map((result, index) => {
+  buildRecordFields(records: GroupRecordEntry[]): EmbedFieldData[] {
+    return records.map((result, index) => {
       const name = result.displayName;
-      const value = toKMB(result.gained);
+      const value = toKMB(result.value);
+      const date = formatDate(result.updatedAt, "DD MMM 'YY");
 
       return {
         name: `${index + 1}. ${name}`,
-        value: `\`${value}\``,
+        value: `\`${value}\` ${date}`,
         inline: true
       };
     });
   }
 }
 
-export default new GainedCommand();
+export default new RecordsCommand();
