+++ src/utils/index.ts
@@ -2,5 +2,4 @@ export * from './dates';
 export * from './discord';
 export * from './levels';
 export * from './metrics';
-export * from './snapshots';
 export * from './strings';
